// src/sdk/database/orm/orm_users.ts
export enum UsersRole {
  Unspecified = 0,
  FREE = 1,
  PAID_ADMIN = 2,
  DEVELOPER = 3,
}
// ...
export interface UsersModel {
  id: string;
  data_creator: string;
  data_updater: string;
  create_time: string;
  update_time: string;
  full_name: string;
  email: string;
  password_hash: string;
  phone?: string | null;
  role: UsersRole;
  organization_id?: string | null;
  last_login?: string | null;
  is_active: boolean;
}